+++ iModelBridgeCore/Pointools/PointoolsVortexAPI/include/ptds/GraphManager.h
@@ -838,7 +838,7 @@ namespace ptds
 
 		p3 = p2 + delta;
 
-		p3[1] = (getExtentsMax() - getExtentsMin())[1] * static_cast<double>(rand()) / static_cast<double>(RAND_MAX);
+		p3[1] = static_cast<float>((getExtentsMax() - getExtentsMin())[1] * static_cast<double>(rand()) / static_cast<double>(RAND_MAX));
 
 		add(p3);
 	}
@@ -1014,7 +1014,7 @@ namespace ptds
 			p += get(i);
 		}
 
-		p *= 1.0 / static_cast<T>(numPoints);
+		p *= 1.0f / static_cast<T>(numPoints);
 
 		return p;
 	}
@@ -1041,7 +1041,7 @@ namespace ptds
 			p += get(i);
 		}
 
-		p *= 1.0 / static_cast<T>(numPoints);
+		p *= 1.0f / static_cast<T>(numPoints);
 
 		return p;
 	}
