+++ iModelBridgeCore/Pointools/PointoolsVortexAPI/src/ptds/DataSourceStructuredStorage.cpp
@@ -180,7 +180,7 @@ DataSource::Size DataSourceStructuredStorage::readBytes(Data *buffer, Size numbe
 
 		beginRead(numberByesToRead);
 
-		if(FAILED(hres = getStream()->Read(buffer, numberByesToRead, &numberBytesRead)))
+		if(FAILED(hres = getStream()->Read(buffer, static_cast<ULONG>(numberByesToRead), &numberBytesRead)))
 			return 0;
 
 		endRead(numberByesToRead);
@@ -196,7 +196,7 @@ DataSource::Size DataSourceStructuredStorage::writeBytes(const Data *buffer, Siz
 
 	if(validHandle())
 	{
-		if(FAILED(getStream()->Write(buffer, numberBytesToWrite, &numberBytesWritten)))
+		if(FAILED(getStream()->Write(buffer, static_cast<ULONG>(numberBytesToWrite), &numberBytesWritten)))
 			return 0;
 	}
