+++ iModelBridgeCore/Pointools/PointoolsVortexAPI/src/ptengine/pointBoundsTree.cpp
@@ -255,12 +255,12 @@ bool PointsBoundsTree::extendTree(
 	
 	tree->pointsProcessed = 0;
 
-	float vol_region = region.extents().x * region.extents().y * region.extents().z;
+    float vol_region = static_cast<float>(region.extents().x * region.extents().y * region.extents().z);
 	if (vol_region < 0.001) vol_region = 1.0;
 
-	float vol_cloud = cloud->root()->extents().volume();
+	float vol_cloud = static_cast<float>(cloud->root()->extents().volume());
 
-	tree->totalPoints = cloud->numPoints() * vol_cloud / vol_region;	// of course this is not exact
+    tree->totalPoints = static_cast<int64_t>(cloud->numPoints() * vol_cloud / vol_region);	// of course this is not exact
 
 	// extract overlapping tree nodes
 	std::set< const TreeNode* > leaves;
